requirements.txt
torch>=2.0.0
torchvision>=0.15.0

numpy
pillow

onnx
onnxscript
onnxruntime

matplotlib
scikit-learn
